Notification JUnit. Creating notifications from events doesn't work properly cause the Event mechanic is not fully adapted on the core.
git-svn-id: https://shanks-with-mason.googlecode.com/svn/trunk@635 6c4d5907-d6a1-e102-47d4-e4e6a2087a27

+++ shanks-core/src/main/java/es/upm/dit/gsi/shanks/model/scenario/Scenario.java
@@ -472,7 +472,7 @@ public abstract class Scenario {
         while (it.hasNext()) {
             Class<? extends Event> type = it.next();
             double prob = 0;
-            Constructor<? extends Event> c = type.getConstructor(new Class[] {String.class, Steppable.class, Double.class});
+            Constructor<? extends Event> c = type.getConstructor(new Class[] {String.class, Steppable.class, Integer.class});
             Event event = c.newInstance(sim.getScenarioManager());
             if(event instanceof ProbabilisticNetworkElementEvent){
                 List<Set<NetworkElement>> list = this.getPossibleEventsOfNE()
